fix small mistake in multiboot header

+++ sys/include/multiboot2.h
@@ -232,7 +232,7 @@ struct multiboot2_tag_framebuffer_common {
     uint32_t type;
     uint32_t size;
 
-    uint32_t framebuffer_addr;
+    uint64_t framebuffer_addr;
     uint32_t framebuffer_pitch;
     uint32_t framebuffer_width;
     uint32_t framebuffer_height;
